requirements.txt
haystack-ai==2.18.0
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.6
